requirements.txt
pyrogram>=2.0.0
tgcrypto  # For better performance with Pyrogram
google-generativeai>=0.3.0
anthropic>=0.8.0
aiohttp>=3.9.0
python-dotenv>=0.19.0
pillow>=9.0.0  # For image handling
python-magic  # For mime type detection
python-docx>=0.8.11  # For Word document processing
PyPDF2>=3.0.0  # For PDF processing
fpdf2==2.7.6
python-docx==1.0.0
markdown==3.5.1
